# openapi.yaml
openapi: 3.1.0
info:
  title: Jywa API
  version: 1.0.0
  description: |
    Fondasi API untuk platform PPOB Jywa. Saat ini mencakup manajemen user, role,
    JWT authentication, saldo, dan audit mutasi saldo.

    Otentikasi: Bearer JWT didapat dari `POST /api/auth/register` atau
    `POST /api/auth/login`.
  contact:
    name: Jywa
    email: admin@jywa.tech
  license:
    name: Private
servers:
  - url: https://jywa.tech
    description: Production
  - url: http://127.0.0.1:3000
    description: Local development
...
paths:
...
  /api/auth/register:
    post:
      operationId: register
      tags: [Auth]
      summary: Registrasi user baru
      security: []
      description: |
        Membuat user baru dengan role default `bronze`.
        Endpoint ini rate-limited.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              basic:
                value:
                  username: user1
                  email: user1@example.com
                  phone: '08123456789'
                  password: password-123
      responses:
        '201':
          description: User berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/auth/login:
    post:
      operationId: login
      tags: [Auth]
      summary: Login
      security: []
      description: |
        Login memakai `identifier` berisi username atau email.
        Endpoint ini rate-limited.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              username:
                value:
                  identifier: jayadana
                  password: password-123
              email:
                value:
                  identifier: user@example.com
                  password: password-123
      responses:
        '200':
          description: Login berhasil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /api/users:
    get:
      operationId: listUsers
      tags: [Owner]
      summary: Daftar semua user
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: role
          in: query
          schema:
            type: string
            enum: [owner, bronze, silver, gold, reseller]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, banned]
        - name: search
          in: query
          schema:
            type: string
          description: Cari di username, email, atau phone
      responses:
        '200':
          description: Daftar user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/digiflazz/webhook:
    post:
      operationId: receiveDigiflazzWebhook
      tags: [Digiflazz]
      summary: Terima webhook Digiflazz
      description: Endpoint publik yang memverifikasi X-Hub-Signature dengan HMAC-SHA1.
      security: []
      parameters:
        - name: X-Hub-Signature
          in: header
          required: true
          schema:
            type: string
            example: sha1=...
        - name: X-Digiflazz-Event
          in: header
          schema:
            type: string
            example: update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data:
                  type: object
                  additionalProperties: true
      responses:
        '200':
          description: Webhook diterima
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: Webhook secret belum dikonfigurasi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/stats:
    get:
      operationId: getV1Stats
      tags: [API v1]
      summary: Statistik penjualan dan profit
      description: |-
        Statistik global untuk monitoring owner. Wajib memakai API key milik user
        dengan role owner. Nilai penjualan, modal, dan profit hanya menghitung
        transaksi berstatus success. Bisa difilter memakai tanggal inklusif.
      security:
        - apiKeyAuth: []
      parameters:
        - name: from
          in: query
          description: Tanggal mulai inklusif (YYYY-MM-DD)
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
        - name: to
          in: query
          description: Tanggal akhir inklusif (YYYY-MM-DD)
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
      responses:
        '200':
          description: Ringkasan statistik
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/stats:
    get:
      operationId: getLegacyStats
      tags: [Owner]
      summary: Statistik penjualan via JWT
      description: Alias internal owner-only untuk dashboard lama.
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
        - name: to
          in: query
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
      responses:
        '200':
          description: Ringkasan statistik
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: API key reseller pada header Authorization.
...
  responses:
    ValidationError:
      description: Request tidak valid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: ValidationError
            message: 'username must be 3-32 chars'
    Unauthorized:
      description: Token hilang, invalid, atau expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unauthorized
            message: Missing bearer token
    Forbidden:
      description: User tidak punya akses atau akun tidak aktif
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Forbidden
            message: Insufficient role
...
    Conflict:
      description: Konflik data unik atau saldo tidak mencukupi
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Conflict
            message: username or email is already registered
    TooManyRequests:
      description: Rate limit tercapai
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: TooManyRequests
            message: too many authentication attempts; try again later
...
  schemas:
...
    StatsResponse:
      type: object
      required: [total_transactions, successful_transactions, failed_transactions, pending_transactions, total_sales, total_cost, total_profit, top_products]
      properties:
        total_transactions:
          type: integer
        successful_transactions:
          type: integer
        failed_transactions:
          type: integer
        pending_transactions:
          type: integer
        total_sales:
          type: integer
          description: Total harga jual transaksi sukses (rupiah).
        total_cost:
          type: integer
          description: Total harga modal transaksi sukses (rupiah).
        total_profit:
          type: integer
          description: Total sales dikurangi total cost (rupiah).
        top_products:
          type: array
          items:
            type: object
            required: [sku, sold, revenue, cost, profit]
            properties:
              sku:
                type: string
              sold:
                type: integer
              revenue:
                type: integer
              cost:
                type: integer
              profit:
                type: integer
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
      required: [error, message]
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        role:
          type: string
          enum: [owner, bronze, silver, gold, reseller]
        balance:
          type: integer
          description: Saldo dalam Rupiah
        status:
          type: string
          enum: [active, suspended, banned]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required: [username, email, phone, password]
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [password]
      properties:
        identifier:
          type: string
          description: username atau email
        username:
          type: string
          description: fallback kompatibilitas
        password:
          type: string
...
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
...
    UserListResponse:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'